third_party/blink/web_tests/printing/no-content-empty-pages-expected.txt

Test for no-content empty pages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS: page number of "page1-1" is 1
PASS: page number of "page1-last" is 1
All tests passed

PASS successfullyParsed is true

TEST COMPLETE
